backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pymongo
python-dotenv
motor
fastapi-cors
email-validator
python-decouple
opencv-python
tensorflow
keras
numpy
pillow
pytest
pytest-cov
requests
mediapipe==0.10.0
scikit-learn==1.2.2
pandas==2.0.0
opencv-python==4.7.0.72
